Refactoring of multipart MIME handling code
Moved all the multipart-related code out of the db package
(attachment.go) into the rest package (multipart.go, a new file.)

Optimized writing nested multipart documents: the nested document no
longer gets written to a temporary in-memory buffer first.
